moved event click function into utils file
