+++ test-result-summary-client/src/Dashboard/Widgets/Graph/utils.js
@@ -1,4 +1,5 @@
 import math from 'mathjs';
+import Swal from 'sweetalert2';
 import BenchmarkMath from '../../../PerfCompare/lib/BenchmarkMath';
 
 export const parseSha = (str, sha) => {
@@ -56,4 +57,41 @@ export const getStatisticValues = (resultsByJDKBuild, key) => {
 
     });
     return [data, std, mean, median];
+}
+
+export const handlePointClick = (event) => {
+    const { buildName, buildNum, javaVersion, jdkDate, testId } = event.point.additionalData[0];
+
+    const buildLinks = ` <a href="/output/test?id=${testId}">${buildName} #${buildNum}</a>`;
+    const CIstr = (typeof event.point.CI === 'undefined') ? `` : `CI = ${event.point.CI}`;
+
+    let ret = `<b>${'NAME'}:</b> ${event.y}<br/> <b>Build: </b> ${jdkDate} <pre>${javaVersion}</pre><br/><b>Link to builds:</b> ${buildLinks}<br /> ${CIstr}`;
+    
+    let i = event.point.series.data.indexOf(event.point);
+    let prevPoint = i === 0 ? null : event.point.series.data[i - 1];
+    let lengthPrev = prevPoint ?
+    prevPoint.additionalData.length : 0;
+    let prevJavaVersion = prevPoint ? prevPoint.additionalData[lengthPrev - 1].javaVersion : null;
+
+    prevJavaVersion = parseSha(prevJavaVersion, 'OpenJ9');
+    javaVersion = parseSha(javaVersion, 'OpenJ9');
+
+    if (prevJavaVersion && javaVersion) {
+        let githubLink = `<a href="https://github.com/eclipse/openj9/compare/${prevJavaVersion}…${javaVersion}">Github Link </a>`;
+        ret += `<br/> <b> Compare Builds: </b>${githubLink}`;
+    }
+
+    Swal.fire({
+        html: ret,
+        showCloseButton: true,
+        showConfirmButton: false,
+        width: '50%',
+        customClass: {
+            htmlContainer: 'text-align: left !important;',
+            container: 'text-align: left !important;',
+              content: 'text-align: left !important;',
+              input: 'text-align: left !important;',
+
+        }
+    });
 }
